pyproject.toml
[tool.poetry]
name = "dashub"
version = "0.1.0"
description = ""
authors = ["Sebastien Campion"]

[tool.poetry.dependencies]
python = "^3.9"
aioredis = "^2.0.1"
starlette = "^0.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
